requirements.txt
openpyxl==2.4.0
requests==2.6.0
lxml==3.6.4
numpy==1.11.1
google-api-python-client==1.5.3
scipy==0.17.0
pdfminer.six==20160614